requirements.txt
streamlit
torch
transformers
sentence-transformers
PyMuPDF
faiss-cpu
